requirements.txt
python-dotenv
requests
pandas
openpyxl
rapidfuzz
unidecode